% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{astrocyte_metadata}
\alias{astrocyte_metadata}
\title{Astrocyte dataset metadata}
\format{
A data frame with 36 rows (samples) and 3 columns (metadata)
}
\usage{
data(astrocyte_metadata)
}
\description{
The metadata for the astrocyte Ribotag dataset (Itoh et al. PNAS. 2018). See ?astrocyte_data for expression data.
}
\details{
#' \itemize{
  \item Sample the sample name corresponding to columns of datExpr
  \item Status experimental autoimmune encephalitis (EAE) or wildtype (WT)
  \item Tissue cerebellum (Cbl), cortex (Ctx), hippocampus (Hippo), or spinal cord (Sc)
}
}
\keyword{datasets}
