% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{astrocyte_networks}
\alias{astrocyte_networks}
\title{Astrocyte networks}
\format{
A list of WGCNA objects (level 1: combined; level 2: EAE, WT; and level 3: Cbl, Ctx, Hippo, Sc)
}
\usage{
data(astrocyte_networks)
}
\description{
The astrocyte networks from Tommasini and Fogel, BMC Bioinformatics, 2023. 
derived from the astrocyte Ribotag data from Itoh et al. PNAS. 2018.
}
\keyword{datasets}
