% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{astrocyte_se}
\alias{astrocyte_se}
\title{Astrocyte Ribotag Summarized Experiment}
\format{
An SummarizedExperiment object with 20352 genes and 36 samples
}
\usage{
data(astrocyte_se)
}
\description{
The astrocyte ribotag dataset from Itoh et al. PNAS. 2018, extracted from the supplementary materials, 
provided as a SummarizedExperiment object. See ?astrocyte_networks for pre-computed WGCNA networks.
}
\keyword{datasets}
