% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{autism_metadata}
\alias{autism_metadata}
\title{Autism dataset metadata}
\format{
A data frame with 58 rows (samples) and 3 columns (metadata)
}
\usage{
data(autism_metadata)
}
\description{
The metadata for the autism dataset (Voineagu et al. 2011). See ?datExpr for expression data.
}
\details{
#' \itemize{
  \item Sample the sample name corresponding to columns of datExpr
  \item Status autism or control
  \item Tissue frontal cortex (FC) or temporal cortex (TC)
}
}
\keyword{datasets}
