% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{autism_se}
\alias{autism_se}
\title{Autism Summarized Experiment}
\format{
An expression data frame with 9934 rows (genes) and 59 columns (probe names + 58 samples)
}
\usage{
data(autism_se)
}
\description{
The autism dataset from Voineagu et al. 2011, extracted from the Gene Expression Omnibus (accession number GSE28521). 
See ?sampleTable for metadata.
}
\keyword{datasets}
