% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{BuildTOMFlowDF}
\alias{BuildTOMFlowDF}
\title{BuildTOMFlowDF}
\usage{
BuildTOMFlowDF(WGCNAlist, networks, toms, genes_to_label, method = "average")
}
\arguments{
\item{WGCNAlist}{list of WGCNA objects}

\item{networks}{list of network names of length 2}

\item{toms}{a list of TOM distance objects of length 2}

\item{genes_to_label}{genes to label across two networks}

\item{method}{linkage method to pass to flashClust for clustering, default is average linkage}
}
\value{
a data.frame
}
\description{
Preprocess for plotting a sankey flow diagram showing the movement of genes from one WGCNA
to another WGCNA. Uses the flashClust framework.
}
\author{
Dario Tommasini, Xinye Li
}
