% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{cleanDatExpr}
\alias{cleanDatExpr}
\title{cleanDatExpr}
\usage{
cleanDatExpr(datExpr, checkGenesSamples = F)
}
\arguments{
\item{datExpr}{a data.frame were columns are samples and rows are samples and the gene symbols are in the first row}

\item{checkGenesSamples}{call the WGCNA function checkGenesSamples?}
}
\description{
A function that converts a data.frame where row 1 is gene symbols to a 
numeric matrix where columns are genes and rows are samples for compatibility 
with most WGCNA functions.
}
\examples{

datExpr=data.frame(c("A", "B", "C"), matrix(sample.int(100, size=12, replace=TRUE),nrow=3,ncol=4), c("mod_1", "mod_2", "mod_1"))
colnames(datExpr)=c("X", "sample.1", "sample.2", "sample.3", "sample.4", "dynamicLabels")
cleanDatExpr(datExpr)

}
\author{
Dario Tommasini
}
