% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{computeOverlapsFromWGCNA}
\alias{computeOverlapsFromWGCNA}
\title{computeOverlapsFromWGCNA}
\usage{
computeOverlapsFromWGCNA(dataset1, dataset2)
}
\arguments{
\item{dataset1}{an object of class WGCNA to compare with dataset2}

\item{dataset2}{an object of class WGCNA to compare with dataset1}
}
\description{
Computes overlap between the modules of two objects of class WGCNA
}
\author{
Dario Tommasini
}
