% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationTests.R
\name{diseasePreservationPtest}
\alias{diseasePreservationPtest}
\title{Disease preservation permutation test}
\usage{
diseasePreservationPtest(
  referenceDatExpr,
  design,
  nPermutations = 20,
  refColumn = 3,
  testColumn = 2,
  nCores = 1,
  saveRData = T
)
}
\arguments{
\item{referenceDatExpr}{data.frame containing expression levels}

\item{design}{a data.frame with the columns corresponding to sample traits of referenceDatExpr}

\item{nPermutations}{number of permutations to perform}

\item{refColumn}{column of design table with reference traits}

\item{testColumn}{column of design table with test traits}

\item{nCores}{number of cores to use, default is 1}

\item{saveRData}{save objects to an RData file?}
}
\description{
Tests whether the disease
}
\author{
Dario Tommasini
}
