% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{drawMultiWGCNAnetwork2}
\alias{drawMultiWGCNAnetwork2}
\title{Draw multiWGCNA network (v2)}
\usage{
drawMultiWGCNAnetwork2(
  WGCNAlist,
  comparisonList,
  moduleOfInterest,
  design,
  removeOutliers = TRUE,
  alpha = 1e-50,
  alpha.vertex = 1e-10,
  vertex.label.cex = 0.5,
  vertex.label.degree = 0,
  vertex.label.dist = 1,
  vertex.label.color = "red3",
  layout = NULL,
  hjust = 0.4,
  vjust = 0.3,
  width = 0.5,
  colors = NULL
)
}
\arguments{
\item{WGCNAlist}{list of WGCNA objects}

\item{comparisonList}{the list of overlap comparisons ie from iterate(myNetworks, overlapComparisons, ...)}

\item{moduleOfInterest}{module of interest, ie "combined_001"}

\item{design}{the sampleTable design matrix}

\item{removeOutliers}{remove outlier modules?}

\item{alpha}{alpha level of edge significance}

\item{alpha.vertex}{alpha level of edge significance for printing vertex labels}

\item{vertex.label.cex}{The font size for vertex labels.}

\item{vertex.label.degree}{It defines the position of the vertex labels, relative to the center of the vertices. It is interpreted as an angle in radians, zero means ‘to the right’, and ‘pi’ means to the left, up is -pi/2 and down is pi/2.}

\item{vertex.label.dist}{The distance of the label from the center of the vertex. If it is 0 then the label is centered on the vertex. If it is 1 then the label is displayed beside the vertex.}

\item{vertex.label.color}{The color of the vertex (node) label text}

\item{layout}{layout of network to be passed to plot function of igraph object, defaults to multiWGCNA custom layout}

\item{hjust}{horizontal justification of labels}

\item{vjust}{vertical justification of labels}

\item{width}{width of labels}

\item{colors}{colors to use for modules, should be the same length as the number of WGCNA objects in the WGCNAlist. Defaults to random colors for each condition.}
}
\value{
list of significant edges
}
\description{
Draw a network where nodes are modules and edges represent significant gene overlap. 
Modules are sorted by levels 1, 2, and 3. Nodes with edges < `alpha.vertex` are labeled
and the edge list is returned by the function of use by the user.
}
\examples{
library(ExperimentHub)
eh = ExperimentHub()
eh_query = query(eh, c("multiWGCNAdata"))
astrocyte_se = eh_query[["EH8223"]]
sampleTable = colData(astrocyte_se)
astrocyte_networks = eh_query[["EH8222"]]
results = list()
results$overlaps = iterate(astrocyte_networks, overlapComparisons, plot=FALSE)
drawMultiWGCNAnetwork(astrocyte_networks, 
  results$overlaps, 
  "combined_013", 
  sampleTable)
  
}
\author{
Dario Tommasini (Modified Timothy Stephens)
}
