% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{getModule}
\alias{getModule}
\title{name: Name of WGCNAobject}
\usage{
getModule(WGCNAobject, gene, use = "dynamicLabels")
}
\arguments{
\item{WGCNAobject}{an object of class WGCNA}
}
\value{
A module name
}
\description{
Returns the module name for a given gene
}
\examples{
library(ExperimentHub)
eh = ExperimentHub()
eh_query = query(eh, c("multiWGCNAdata"))
astrocyte_networks = eh_query[["EH8222"]]
getModule(astrocyte_networks[[1]], 'Srgn')
}
