% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preservation.R
\name{getPreservation}
\alias{getPreservation}
\title{Module preservation analysis}
\usage{
getPreservation(reference, test, nPermutations = 100, write = FALSE)
}
\arguments{
\item{comparison}{a data.frame resulting from a call to computeOverlapsFromWGCNA}

\item{dataset1}{an object of class WGCNAobject to compare with dataset2}

\item{dataset2}{an object of class WGCNAobject to compare with dataset1}
}
\description{
Performs preservation of reference modules in test dataset
}
\examples{
getPreservation(referenceDatExpr, testDatExpr, nPermutations=50)

}
\author{
Dario Tommasini
}
