% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{GetSignificantOverlap}
\alias{GetSignificantOverlap}
\title{Get significant overlap}
\usage{
GetSignificantOverlap(df, p.adj.threshold = 0.05, overlap.threshold = 10)
}
\arguments{
\item{df}{a data.frame resulting from a call to computeOverlapsFromWGCNA}

\item{p.adj.threshold}{an object of class WGCNA to compare with dataset2}

\item{overlap.threshold}{an object of class WGCNA to compare with dataset1}
}
\value{
Returns a data.frame with significant overlaps
}
\description{
A function that subsets an overlap comaprisons data.frame.
}
\author{
Dario Tommasini
}
