% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{iterate}
\alias{iterate}
\title{Iterate function across networks}
\usage{
iterate(WGCNAlist, FUN, plot = FALSE, write = FALSE)
}
\arguments{
\item{WGCNAlist}{a vector of objects of type WGCNAobject}

\item{FUN}{function to iterate, either overlapComparisons or preservationComparisons}

\item{plot}{generate plots?}
}
\description{
A high level function that iterates functions across
a list of WGCNA objects
}
\examples{

iterate(myNetworks, overlapComparisons, plot=TRUE)

}
\author{
Dario Tommasini
}
