% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{makeTraitTable}
\alias{makeTraitTable}
\title{Generate a trait table from a sample table}
\usage{
makeTraitTable(
  inputTable,
  column,
  detectNumbers = get("detectNumbers", envir = parent.frame())
)
}
\arguments{
\item{inputTable}{the sampleTable dataframe}

\item{column}{the column from the sampleTable to use as traits}

\item{detectNumbers}{whether to consider traits with numbers as numerical rather than categorical variables}
}
\description{
Generates a WGCNA-compatible trait table from a sampleTable dataframe
}
