% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{moduleComparisonPlot}
\alias{moduleComparisonPlot}
\title{Plots an expression profile for a module}
\usage{
moduleComparisonPlot(overlapDf, dataset1, dataset2)
}
\arguments{
\item{overlapDf}{a data.frame resulting from a call to computeOverlapsFromWGCNA}

\item{dataset1}{an object of class WGCNAobject to compare with dataset2}

\item{dataset2}{an object of class WGCNAobject to compare with dataset1}
}
\description{
A plotting function that returns a heatmap and barplot for a module
}
\examples{

moduleComparisonPlot(comparisonList[[element]]$overlap, WGCNAlist[[first]], WGCNAlist[[second]])

}
\author{
Dario Tommasini
}
