% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiWGCNA-package.R
\docType{package}
\name{multiWGCNA-package}
\alias{multiWGCNA}
\alias{multiWGCNA-package}
\title{multiWGCNA: multiWGCNA}
\description{
An R package for deeping mining gene co-expression networks in multi-trait expression data. Provides functions for analyzing, comparing, and visualizing WGCNA networks across conditions. multiWGCNA was designed to handle the common case where there are multiple biologically meaningful sample traits, such as disease vs wildtype across development or anatomical region.
}
\author{
\strong{Maintainer}: Dario Tommasini \email{dtommasini0@gmail.com} (\href{https://orcid.org/0000-0002-1214-6547}{ORCID})

Authors:
\itemize{
  \item Brent Fogel [contributor]
}

}
\keyword{internal}
