% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffModuleExpAnalysis.R
\name{performANOVA}
\alias{performANOVA}
\title{Perform ANOVA}
\usage{
performANOVA(
  datExpr,
  testCondition,
  refCondition,
  design = get("design", envir = parent.frame()),
  alphaLevel = get("alphaLevel", envir = parent.frame())
)
}
\arguments{
\item{datExpr}{expression data.frame}

\item{testCondition}{test column in sampleTable}

\item{refCondition}{reference column in sampleTable}

\item{design}{the sampleTable}

\item{alphaLevel}{the significance level}
}
\description{
Test association between module expression to traits using ANOVA
}
