% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{permutationTestResults}
\alias{permutationTestResults}
\title{Permutation test results}
\format{
A list of data.frames containing preservation results for each 
permutation
}
\usage{
data(permutationTestResults)
}
\description{
The results of running the PreservationPermutationTest in the astrocyte
vignette. This is provided since this function is quite slow. Please see
the astrocyte vignette for more details.
}
\keyword{datasets}
