% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{PlotMultiNodesTOMflow}
\alias{PlotMultiNodesTOMflow}
\title{PlotMultiNodesTOMflow}
\usage{
PlotMultiNodesTOMflow(
  TOMDF,
  labels = NULL,
  alpha = 0.1,
  width = 0.05,
  color = "black"
)
}
\arguments{
\item{TOMDF}{created by BuildTOMFlowDF}

\item{labels}{labels to use for the networks, default is Network1, Network2, etc.}

\item{alpha}{alpha of flows}

\item{width}{width of the strata}

\item{color}{color of flows}
}
\value{
a ggplot object
}
\description{
Plots a sankey flow diagram showing the movement of genes from one WGCNA
to multi-WGCNA networks. Uses the ggalluvial framework.
}
\author{
Dario Tommasini, Xinye Li
}
