% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWGCNA.R
\name{plotSoftPowerPicks}
\alias{plotSoftPowerPicks}
\title{plotSoftPowerPicks: Select softpower for full and sub-networks}
\usage{
plotSoftPowerPicks(softpowers, point_size = 5, text_size = 3)
}
\arguments{
\item{softpowers}{soft powers output from `selectSoftPower`}

\item{point_size}{size of points in plot}

\item{text_size}{size of text in plot}
}
\value{
A list of ggplot objects (one per trait) that can be plotted using `wrap_plots`
}
\description{
A high level function that runs softpower selection
}
\examples{
library(ExperimentHub)
eh = ExperimentHub()
eh_query = query(eh, c("multiWGCNAdata"))
autism_se = eh_query[["EH8219"]]
set.seed(1)
autism_se = autism_se[sample(rownames(autism_se), 500),]
sampleTable = colData(autism_se)
conditions1 = unique(sampleTable[,2])
conditions2 = unique(sampleTable[,3])

sp = selectSoftPower(autism_se, sampleTable, conditions1, conditions2, 
  verbose = 5)

plots <- plotSoftPowerPicks(sp)
library(patchwork)
wrap_plots(plots, ncol=2)

}
\author{
Timothy Stephens
}
