% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preservation.R
\name{preservationComparisons}
\alias{preservationComparisons}
\title{Preservation comparisons}
\usage{
preservationComparisons(
  comparisonList,
  WGCNAlist,
  first,
  second,
  element,
  plot = FALSE,
  alpha = get("alpha", envir = parent.frame())
)
}
\arguments{
\item{comparisonList}{a list passed by the iterate function}

\item{WGCNAlist}{list of objects of type WGCNAobject}

\item{first}{index of first WGCNAobject}

\item{second}{index of second WGCNAobject}

\item{element}{element position in the comparison list (passed by iterate function)}

\item{plot}{generate plots?}
}
\description{
A high level function that performs a perservation comparison between two
WGCNAobjects in a WGCNAlist, usually supplied by iterate function
}
\examples{

results$preservation=iterate(myNetworks[conditions1], preservationComparisons, plot=TRUE)

}
\author{
Dario Tommasini
}
