% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWGCNA.R
\name{selectSoftPower}
\alias{selectSoftPower}
\title{selectSoftPower: Select softpower for full and sub-networks}
\usage{
selectSoftPower(
  datExpr,
  sampleTable,
  conditions1,
  conditions2,
  powers = c(seq(from = 1, to = 20, by = 1)),
  write = FALSE,
  alphaLevel = 0.05,
  plot = FALSE,
  ...
)
}
\arguments{
\item{datExpr}{either a SummarizedExperiment object or data.frame with genes are rows and samples as columns}

\item{sampleTable}{data.frame with sample names in first column and sample traits in the second and third column. First column should be called "Sample"}

\item{conditions1}{first design conditions, ie healthy/disease}

\item{conditions2}{second design conditions, ie frontal lobe/temporal lobe}

\item{powers}{list of powers to check}

\item{...}{Arguments to pass to pickSoftThreshold function}
}
\value{
A list of softpower results, ie level one, two, and three networks.
}
\description{
A high level function that runs softpower selection
}
\examples{
library(ExperimentHub)
eh = ExperimentHub()
eh_query = query(eh, c("multiWGCNAdata"))
autism_se = eh_query[["EH8219"]]
set.seed(1)
autism_se = autism_se[sample(rownames(autism_se), 500),]
sampleTable = colData(autism_se)
conditions1 = unique(sampleTable[,2])
conditions2 = unique(sampleTable[,3])

sp = selectSoftPower(autism_se, sampleTable, conditions1, conditions2, 
  verbose = 5)
sp[["combined"]]

}
\author{
Timothy Stephens
}
