% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{summarizeResults}
\alias{summarizeResults}
\title{summarizeResults: Summarize results from a results list object}
\usage{
summarizeResults(
  myNetworks,
  results,
  alphaLevel = 0.05,
  write = FALSE,
  outputFile = "results.txt"
)
}
\arguments{
\item{myNetworks}{a list of WGCNAobjects}

\item{results}{results list}

\item{alphaLevel}{alpha level of significance}

\item{write}{write to file?}

\item{outputFile}{name of output file, defaults to results.txt}
}
\value{
prints a summary of results from the multiWGCNA analysis
}
\description{
Prints (or writes) a summary of the results from a results list object
}
