% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tau_se}
\alias{tau_se}
\title{Tau pathology (rTg4510) Mouse Summarized Experiment}
\format{
A SummarizedExperiment object with 18822 genes and 58 samples
}
\usage{
data(tau_se)
}
\description{
The tau pathology dataset from Castanho et al. Cell Rep. 2020, extracted from GSE125957, 
processed as described in methods (https://git.exeter.ac.uk/ic322/ad-mice-rna-seq-cell-reports), 
and provided as a SummarizedExperiment object. Values correspond to DESeq2 rlog-transformed 
counts. See ?tau_networks for pre-computed WGCNA networks.
}
\keyword{datasets}
