% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{topNGenes}
\alias{topNGenes}
\title{Top N genes of a module}
\usage{
topNGenes(WGCNAobject, module, nGenes = NULL)
}
\arguments{
\item{WGCNAobject}{an object of type WGCNAobject}

\item{module}{the name of the module in WGCNAobject}

\item{nGenes}{an integer from 1 to module size; returns all genes if left NULL}
}
\description{
Returns the top N number of genes of a module.
All genes returned if no number is specified.
Genes are in order of intramodular connectivity.
}
