% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{astrocyte_data}
\alias{astrocyte_data}
\title{Astrocyte dataset}
\format{
An expression data frame with 20352 rows (genes) and 37 columns (gene names + 36 samples)
}
\usage{
data(astrocyte_data)
}
\description{
The astrocyte dataset from Itoh et al. PNAS. 2018, extracted from the supplementary materials.  
See astrocyte_metadata for metadata.
}
\keyword{datasets}
