% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{bidirectionalBestMatches}
\alias{bidirectionalBestMatches}
\title{Best matching modules}
\usage{
bidirectionalBestMatches(overlapDf, plot = TRUE)
}
\arguments{
\item{overlapDf}{a data.frame resulting from a call to computeOverlapsFromWGCNA}

\item{plot}{generate a heatmap of best matching modules?}
}
\description{
Find all the modules from dataset1 that have a best match to a module in dataset2
if that module in dataset2 is also a best match to the module in dataset1
}
\author{
Dario Tommasini
}
