% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWGCNA.R
\name{constructNetworks}
\alias{constructNetworks}
\title{constructNetworks: Construct all the weighted gene correlation networks}
\usage{
constructNetworks(
  datExpr,
  sampleTable,
  conditions1,
  conditions2,
  write = FALSE,
  alphaLevel = 0.05,
  plot = FALSE,
  ...
)
}
\arguments{
\item{datExpr}{data.frame where column 1 has samples and row 1 has genes}

\item{sampleTable}{data.frame with sample traits}

\item{conditions1}{first design conditions, ie healthy/disease}

\item{conditions2}{second design conditions, ie frontal lobe/temporal lobe}

\item{write}{write results out to files?}

\item{alphaLevel}{significance value passed to findBestTrait function, default is 0.05}

\item{plot}{plot modules? Default is false}

\item{...}{Arguments to pass to blockwiseModules function}
}
\description{
A high level function that returns all networks
possible for a given experimental design
}
\author{
Dario Tommasini
}
