% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWGCNA.R
\name{constructNetworks}
\alias{constructNetworks}
\title{Construct all the weighted gene correlation networks}
\usage{
constructNetworks(
  datExpr,
  sampleTable,
  conditions1,
  conditions2,
  write = FALSE,
  alphaLevel = get("alphaLevel", envir = parent.frame()),
  plot = FALSE,
  ...
)
}
\arguments{
\item{datExpr}{data.frame where column 1 has samples and row 1 has genes}

\item{sampleTable}{data.frame with sample traits}

\item{conditions1}{first design conditions, ie healthy/disease}

\item{conditions2}{second design conditions, ie frontal lobe/temporal lobe}

\item{softPower}{thresholding power used for network construction}

\item{originalDirectory}{directory to output files, defaults to getwd()}

\item{blockWise}{use the blockwiseModules function? default is TRUE}

\item{nCores}{number of cores to use for multithreaded network construction}
}
\description{
A high level function that returns all networks
possible for a given experimental design
}
\examples{

myNetworks=constructNetworks(datExpr, sampleTable, conditions1, conditions2, softPower=10, blockWise=T, nCores=nCores)

}
\author{
Dario Tommasini
}
