% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffModuleExpAnalysis.R
\name{diffModuleExpression}
\alias{diffModuleExpression}
\title{Differential module expression}
\usage{
diffModuleExpression(
  WGCNAobject,
  geneList,
  moduleName = NULL,
  mode = "PC1",
  design = sampleTable,
  testColumn = 2,
  refColumn = 3,
  test = "ANOVA",
  plot = TRUE
)
}
\arguments{
\item{WGCNAobject}{WGCNA object}

\item{geneList}{vector of genes in WGCNAobject}

\item{moduleName}{name of the module for the title}

\item{mode}{either PC1 or Zscore, default is PC1}

\item{design}{the sampleTable}

\item{testColumn}{the column of the sampleTable to be resolved}

\item{refColumn}{the column of the sampleTable to be used as biological variation}

\item{test}{statistical test to perform, either "ANOVA" or "PERMANOVA"}

\item{plot}{generate a plot?}
}
\description{
Runs (and plots if turned on) the differential module expression analysis
}
