% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preservation.R
\name{getPreservation}
\alias{getPreservation}
\title{getPreservation}
\usage{
getPreservation(reference, test, nPermutations = 100, write = FALSE)
}
\arguments{
\item{reference}{reference network of class WGCNA}

\item{test}{test network of class WGCNA}

\item{nPermutations}{number of permutations to perform; at least 50 permutations}

\item{write}{write to file?}
}
\value{
a data.frame summarizing results of preservation analysis
}
\description{
Performs a network preservation analysis
}
\author{
Dario Tommasini
}
