% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{makeTraitTable}
\alias{makeTraitTable}
\title{generate a trait table from a design table}
\usage{
makeTraitTable(
  inputTable,
  column,
  detectNumbers = get("detectNumbers", envir = parent.frame())
)
}
\description{
generate a trait table from a design table
}
