% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffModuleExpAnalysis.R
\name{moduleExpressionPlot}
\alias{moduleExpressionPlot}
\title{Plots an expression profile for a module}
\usage{
moduleExpressionPlot(
  WGCNAobject,
  geneList,
  mode = "PC1",
  legend = FALSE,
  title = NULL,
  clusterGenes = F
)
}
\arguments{
\item{WGCNAobject}{an object of class WGCNAobject}

\item{geneList}{a vector of gene names to be extracted from WGCNAobject}

\item{mode="PC1"}{use first principal component or averageZscore?}

\item{legend=FALSE}{plot legend?}

\item{title=NULL}{title of the plot}

\item{clusterGenes=F}{cluster heatmap genes by hierarchical clustering?}
}
\description{
A plotting function that returns a heatmap and barplot for a module
}
\examples{

moduleExpressionPlot(WGCNAobject, moduleGenes, title=myModule)

}
\author{
Dario Tommasini
}
