% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{overlapComparisons}
\alias{overlapComparisons}
\title{Overlap comparisons}
\usage{
overlapComparisons(
  comparisonList,
  WGCNAlist,
  first,
  second,
  element,
  plot = TRUE,
  write = FALSE
)
}
\arguments{
\item{comparisonList}{a list passed by the iterate function}

\item{WGCNAlist}{list of objects of type WGCNAobject}

\item{first}{index of first WGCNAobject}

\item{second}{index of second WGCNAobject}

\item{element}{element position in the comparison list (passed by iterate function)}

\item{plot}{generate plots?}

\item{write}{write results to file?}
}
\description{
Compares modules between two objects of type WGCNAobjects
within a WGCNAobject list given the indices
}
\author{
Dario Tommasini
}
