% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preservation.R
\name{preservationComparisonPlot}
\alias{preservationComparisonPlot}
\title{Preservation scatterplot}
\usage{
preservationComparisonPlot(
  comparison,
  dataset1,
  dataset2,
  alphaLevel,
  outliers = FALSE
)
}
\arguments{
\item{comparison}{a data.frame resulting from a call to computeOverlapsFromWGCNA}

\item{dataset1}{an object of class WGCNAobject to compare with dataset2}

\item{dataset2}{an object of class WGCNAobject to compare with dataset1}

\item{alphaLevel}{alpha level of significance}

\item{outliers}{leave outlier modules? By default these are removed}
}
\description{
A plotting  function that draws a scatterplot of preservation scores between
two WGCNAobjects
}
\examples{

preservationComparisonPlot(comparisonList[[1]], WGCNAlist[[1]], WGCNAlist[[2]], 0.05)

}
\author{
Dario Tommasini
}
