% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preservation.R
\name{PreservationPermutationTest}
\alias{PreservationPermutationTest}
\title{PreservationPermutationTest}
\usage{
PreservationPermutationTest(
  referenceDatExpr,
  design,
  nPermutations = 100,
  refColumn = 3,
  testColumn = 2,
  nCores = 20,
  nPresPermutations = 50,
  ...
)
}
\arguments{
\item{datExpr}{a data.frame containing expression values}

\item{conditions}{a vector containing conditions for the samples}

\item{geneList}{vector of genes, will use all genes if NULL (default)}

\item{plot}{plot a network?}
}
\value{
A list including a matrix of z-scores, a matrix of raw p-values, a 
matrix of adjusted p-values, and a summary data.frame
}
\description{
Performs a permutation test to determine if a null distribution of expected
preservation scores for modules in this dataset if the labels were 
randomly assigned.
}
\examples{
library(ExperimentHub)
eh = ExperimentHub()
eh_query = query(eh, c("multiWGCNAdata"))
astrocyte_se = eh_query[["EH8223"]]
datExpr = assays(astrocyte_se)[[1]]

}
\author{
Dario Tommasini
}
