% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{run_goseq}
\alias{run_goseq}
\title{Run goseq gene ontology}
\usage{
run_goseq(WGCNAobject, module, organism = "mm10", id = "geneSymbol")
}
\description{
Performs a goseq gene ontology analysis of a module
}
\author{
Dario Tommasini
}
