% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffModuleExpAnalysis.R
\name{runDME}
\alias{runDME}
\title{Run differential module expression}
\usage{
runDME(
  WGCNAobject,
  alphalevel = get("alphaLevel", envir = parent.frame()),
  design = sampleTable,
  testCondition = NULL,
  refCondition = NULL,
  p.adjust = "fdr",
  plot = FALSE,
  test = "ANOVA",
  write = FALSE,
  out = NULL
)
}
\arguments{
\item{WGCNAobject}{object of class WGCNA with the modules to run DME on}

\item{design}{the sampleTable}

\item{p.adjust}{adjust for multiple comparisons, argument to pass to p.adjust function}

\item{plot}{generate a plot?}

\item{test}{statistical test to perform, either "ANOVA" or "PERMANOVA"}

\item{write}{write results to a file?}

\item{alphaLevel}{level of significance}

\item{testColumn}{the column of the sampleTable to be resolved}

\item{refColumn}{the column of the sampleTable to be used as biological variation}
}
\description{
A wrapper to run diffModuleExpression on all the modules in a network
}
\examples{

runDME(myNetworks$combined, alpha=0.05)

}
\author{
Dario Tommasini
}
