% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffModuleExpAnalysis.R
\name{runDME}
\alias{runDME}
\title{Best matching modules}
\usage{
runDME(
  WGCNAobject,
  alpha = get("alpha", envir = parent.frame()),
  design = sampleTable,
  testCondition = NULL,
  refCondition = NULL,
  p.adjust = "fdr",
  plot = FALSE,
  write = FALSE
)
}
\arguments{
\item{plot}{generate a heatmap of best matching modules?}

\item{overlapDf}{a data.frame resulting from a call to computeOverlapsFromWGCNA}
}
\description{
Find all the modules from dataset1 that have a best match to a module in dataset2
if that module in dataset2 is also a best match to the module in dataset1
}
\examples{

bidirectionalBestMatches(comparisonList[[element]]$overlap, WGCNAlist[[first]], WGCNAlist[[second]])

}
\author{
Dario Tommasini
}
